Split apart a bunch of code
