+++ server/src/db.ts
@@ -20,19 +20,7 @@ along with Home Lights.  If not, see <http://www.gnu.org/licenses/>.
 import { existsSync, mkdirSync } from 'fs';
 import { join, dirname } from 'path';
 import { getEnvironmentVariable } from './util';
-import { init as initDB, dbRun, dbAll } from './sqlite';
-import {
-  Zone,
-  CreateZoneRequest,
-  LightType,
-  Light,
-  CreateLightRequest,
-  RVLLight,
-  CreateRVLLightRequest,
-  HueLight,
-  CreatePhilipsHueLightRequest
-} from './common/types';
-import { NUM_RVL_CHANNELS } from './common/config';
+import { init as initDB, dbRun } from './sqlite';
 
 const DB_FILE = join(
   getEnvironmentVariable('HOME'),
@@ -81,113 +69,4 @@ export async function init(): Promise<void> {
     await dbRun(PHILIPS_HUE_INFO_SCHEMA);
   }
   console.log('Database initialized');
-}
-
-// ---- Zone Operations ----
-
-export async function getZones(): Promise<Zone[]> {
-  return dbAll(`SELECT * FROM zones`) as Promise<Zone[]>;
-}
-
-export async function createZone(
-  zoneRequest: CreateZoneRequest
-): Promise<void> {
-  await dbRun(`INSERT INTO zones (name) values (?)`, [zoneRequest.name]);
-}
-
-export async function editZone(zone: Zone): Promise<void> {
-  await dbRun('UPDATE zones SET name = ? WHERE id = ?', [zone.name, zone.id]);
-}
-
-export async function deleteZone(id: number): Promise<void> {
-  await dbRun('DELETE FROM zones WHERE id = ?', [id]);
-}
-
-// ---- Light Operations ----
-
-export async function getLights(): Promise<Light[]> {
-  return dbAll(`SELECT * FROM lights`) as Promise<Light[]>;
-}
-
-export async function createLight(
-  lightRequest: CreateLightRequest
-): Promise<void> {
-  switch (lightRequest.type) {
-    case LightType.RVL: {
-      const rvlLightRequest: CreateRVLLightRequest = lightRequest as CreateRVLLightRequest;
-      if (
-        !Number.isInteger(rvlLightRequest.channel) ||
-        rvlLightRequest.channel < 0 ||
-        rvlLightRequest.channel >= NUM_RVL_CHANNELS
-      ) {
-        throw new Error(`Invalid RVL channel ${rvlLightRequest.channel}`);
-      }
-      await dbRun(`INSERT INTO lights (name, type, channel) values (?, ?, ?)`, [
-        rvlLightRequest.name,
-        LightType.RVL,
-        rvlLightRequest.channel
-      ]);
-      break;
-    }
-    case LightType.PhilipsHue: {
-      const philipsHueLightRequest: CreatePhilipsHueLightRequest = lightRequest as CreatePhilipsHueLightRequest;
-      await dbRun(`INSERT INTO lights (name, type) values (?, ?)`, [
-        philipsHueLightRequest.name,
-        LightType.RVL
-      ]);
-      break;
-    }
-  }
-}
-
-export async function editLight(light: Light): Promise<void> {
-  switch (light.type) {
-    case LightType.RVL:
-      const rvlLight: RVLLight = light as RVLLight;
-      await dbRun('UPDATE lights SET name = ?, channel = ? WHERE id = ?', [
-        rvlLight.name,
-        rvlLight.channel,
-        rvlLight.id
-      ]);
-      break;
-    case LightType.PhilipsHue:
-      const hueLight: HueLight = light as HueLight;
-      await dbRun('UPDATE lights SET name = ?, WHERE id = ?', [
-        hueLight.name,
-        hueLight.id
-      ]);
-      break;
-  }
-}
-
-export async function deleteLight(id: number): Promise<void> {
-  await dbRun('DELETE FROM lights WHERE id = ?', [id]);
-}
-
-// ---- Philips Hue Operations ----
-
-export interface PhilipsHueInfo {
-  username: string;
-  key: string;
-}
-
-export async function getPhilipsHueInfo(): Promise<PhilipsHueInfo | null> {
-  const rows = await dbAll(`SELECT * FROM "philips_hue_info"`);
-  switch (rows.length) {
-    case 0:
-      return null;
-    case 1:
-      return rows[0] as PhilipsHueInfo;
-    default:
-      throw new Error(
-        `Internal Error: philips_hue_info unexpectedly has more than one row`
-      );
-  }
-}
-
-export async function setPhilipsHueInfo(info: PhilipsHueInfo): Promise<void> {
-  await dbRun(`INSERT INTO philips_hue_info (username, key) VALUES(?, ?)`, [
-    info.username,
-    info.key
-  ]);
 }
